backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
pymysql==1.1.0
SQLAlchemy==2.0.25
cryptography==41.0.7
python-dotenv==1.0.0
bcrypt==4.1.2
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
yfinance==0.2.36
groq==0.4.2
gunicorn==21.2.0
pandas==2.2.0
numpy==1.26.4